Delete planner list spec
